Fixing Issue #51
