+++ frontend/rievent-frontend/src/styles/CustomTextField.js
@@ -1,0 +1,17 @@
+import { TextField } from '@mui/material/';
+import { styled } from '@mui/styles';
+
+const CustomTextField = styled(TextField)({
+    '& .MuiOutlinedInput-root': {
+        '& fieldset': {
+            border: 'none',
+            textAlign: 'center'
+        },
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        paddingTop: '2%',
+    },
+});
+
+export default CustomTextField;
